requirements.txt
streamlit
google-genai
pypdf
tenacity